refactor(server): title-generation owns persistence
Move prompt builders and persistence behind the title-generation seam:
- buildChatTitleContent from chat/service.ts
- buildRecordingTitleContent from recordings/analysis-service.ts
- chat and recording title persistence (save, cost, emit) from adapters/title-generation.ts

Adapters/title-generation becomes a thin bus adapter delegating to
title-generation/service. Gives locality for title behaviour and makes
the adapter's deletion test pass (two producers justify the seam).
